Index.html and directory of recipe documents
The index.html document was created to be the main document that the website starts with and is a viewpoint for all of the recipe documents.
In the recipes directory will go all of the recipe documents. I started with one of my favorite homemade recipes, a rice-bowl.
More recipes will soon be created and linked to index.html along with their boilplate code.