Added a new recipe and made minor changes in other pages
I added a new recipe to the recipe site where I stated the description, ingredients, and steps in making a burger.
The recipe was detailed in a way so that it was flexible for the person making the dish to use any ingredients that were available to them.
Minor changes were made to other pages including punctuation and adding links to the newly added recipes in index.html.
One more recipe will soon follow.