Added a new recipe and link
I created another new recipe for pasta which includes the same style as the other recipes.
The recipe, just like the others, consists of a description, an image of the dish, the list of ingredients and a list of steps to make the dish.
I am finished with the basic structure of the web page. Later, I will style the page with CSS.