Added styles to the website
I used the command line to create a new directory and touch a stylesheet to use for the website.
Because there were many HTML documents to work with, I thought that I should use the external method to apply styles to all of them.
With the externally linked CSS stylesheet, I applied compatiable styles to the lists, headings and paragraphs.
These styles included a background color, text color and a web-safe font-family.